requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.1.0
streamlit==1.44.1
streamlit-authenticator==0.2.3

